Добавил таблицу Offers. Добавил сервис, который её заполняет.

+++ db/migrate/sql/001_up.sql
@@ -4,7 +4,7 @@ CREATE TABLE categories(
   title       VARCHAR(255),
   updated_at  TIMESTAMP     DEFAULT now(),
   created_at  TIMESTAMP     DEFAULT now(),
-  shop_id     BIGINT UNSIGNED NOT NULL,
+  shop_id     SMALLINT UNSIGNED NOT NULL,
   UNIQUE KEY shop_and_category (id, shop_id),
   PRIMARY KEY (id)
 );
